odin timings; rust benchmark

+++ benchmark/odin/run_bench.py
@@ -15,6 +15,6 @@ for _ in tqdm(range(100)):
 df=pd.DataFrame(timings)
 df.time=df.time.str.strip()
 df['magnitude']=df.time.str.replace(r'[0-9]*\.?[0-9]*', '', regex=True)
-df.time=df.time.str.replace(r'[^\d\.]', '', regex=True)
+df.time=df.time.str.replace(r'[^\d\.]', '', regex=True).astype(float)
 df.time/=df.magnitude.map(mappings)
 df.to_csv('timings.csv', index=False)
